feat(core): extract IngestionService and MaintenanceService (phase-3 PR2)
Pull all LLM orchestration, data lifecycle, and background task logic
out of WikiMemory into dedicated domain services:

- IngestionService owns the ingestDocument pipeline (chunking, LLM calls,
  DB commit, embedding side-effects)
- MaintenanceService owns runPrune, runLibrarian, runHeal, runReembed,
  and forget; absorbs _doRunLibrarian and _doRunHeal implementations
- WikiMemory becomes a thin facade: constructor wires up both services,
  all delegated methods are one-liners; write() auto-trigger routes
  through maintenanceService internals

Includes PR1 prerequisites (pure.ts, JobManager) since this branch
diverged from phase-3. Target PR against phase-3-pr1.
